extracted getParamsFromURL so that it's only once in the code
